# openapi/cloudcommander-v1.yaml
openapi: 3.0.3
info:
  title: CloudCommander API
  version: 1.1.0
  description: Command/query API for CloudCommander deterministic event core.
servers:
  - url: https://api.cloudcommander.example.com
security:
  - bearerAuth: []

paths:
  /api/v1/commands/resource-allocation:
    post:
      summary: Request a node resource allocation change
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XRole'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceAllocationCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Invalid authentication or role claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen or locked by active compensation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/commands/dependency-edge:
    post:
      summary: Propose a dependency edge
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XRole'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DependencyEdgeCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Invalid authentication or role claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen or locked by active compensation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/commands/rollback:
    post:
      summary: Initiate rollback for a target aggregate to a target sequence
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XRole'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Invalid authentication or role claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen or locked by active compensation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/projections/service-graph:
    get:
      summary: Return current projected service graph for tenant
      parameters:
        - $ref: '#/components/parameters/XTenantId'
      responses:
        '200':
          description: Projected service graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceGraphProjection'

  /api/v1/projections/nodes/{node_id}:
    get:
      summary: Return projected node state
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: path
          name: node_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Node projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNodeProjection'
        '404':
          description: Projection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/telemetry/system/backpressure:
    get:
      summary: Return write-path backpressure telemetry
      responses:
        '200':
          description: Backpressure state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackpressureTelemetry'

  /api/v1/telemetry/system/reconciler:
    get:
      summary: Return reconciler circuit breaker telemetry
      responses:
        '200':
          description: Reconciler circuit breaker state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconcilerTelemetry'

  /api/v1/telemetry/graph/centrality:
    get:
      summary: Return graph blast-radius centrality scores
      parameters:
        - $ref: '#/components/parameters/XTenantId'
      responses:
        '200':
          description: Ranked graph centrality scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GraphCentralityNode'

  /api/v1/telemetry/nodes/{node_id}/guardrail-state:
    get:
      summary: Return latest guardrail alert for node
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: path
          name: node_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Guardrail telemetry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailState'

  /api/v1/telemetry/events:
    get:
      summary: Return normalized control-plane telemetry events
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - in: query
          name: source
          required: false
          schema:
            $ref: '#/components/schemas/EventSource'
        - in: query
          name: severity
          required: false
          schema:
            $ref: '#/components/schemas/EventSeverity'
      responses:
        '200':
          description: Normalized telemetry events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryEvent'

  /api/v1/telemetry/events/recent:
    get:
      summary: Return latest control-loop events
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Recent events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecentControllerEvent'

  /api/v1/telemetry/quarantine:
    get:
      summary: Return quarantined causal-conflict events
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Quarantine events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuarantineEvent'

  /api/v1/control/quarantine/{quarantine_id}/resolve:
    post:
      summary: Resolve a quarantined causal-conflict event
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XRole'
        - $ref: '#/components/parameters/XActorClaims'
        - in: path
          name: quarantine_id
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuarantineResolveRequest'
      responses:
        '200':
          description: Quarantine event resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuarantineResolveResponse'
        '403':
          description: Tenant admin claim required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Quarantine event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    XTenantId:
      in: header
      name: X-Tenant-Id
      required: true
      schema:
        type: string
        format: uuid
    XActorId:
      in: header
      name: X-Actor-Id
      required: true
      schema:
        type: string
        minLength: 1
    XActorClaims:
      in: header
      name: X-Actor-Claims
      required: false
      schema:
        type: string
    XRole:
      in: header
      name: X-Role
      required: false
      schema:
        type: string
        enum:
          - user
          - system
        default: user
    XIdempotencyKey:
      in: header
      name: X-Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
    XExpectedVersion:
      in: header
      name: X-Expected-Version
      required: true
      schema:
        type: integer
        minimum: 0

  schemas:
    ResourceAllocationCommand:
      type: object
      additionalProperties: false
      required:
        - node_id
        - target_cpu_cores
        - target_memory_gb
        - reason_code
      properties:
        node_id:
          type: string
          format: uuid
        target_cpu_cores:
          type: number
          minimum: 0.1
          maximum: 128.0
        target_memory_gb:
          type: number
          minimum: 0.5
          maximum: 1024.0
        reason_code:
          type: string
          minLength: 1
          maxLength: 50

    DependencyEdgeCommand:
      type: object
      additionalProperties: false
      required:
        - source_node_id
        - target_node_id
      properties:
        source_node_id:
          type: string
          format: uuid
        target_node_id:
          type: string
          format: uuid
        edge_type:
          type: string
          enum: [sync, async]
          default: sync

    RollbackCommand:
      type: object
      additionalProperties: false
      required:
        - target_aggregate_id
        - target_sequence_id
        - reason_code
      properties:
        target_aggregate_id:
          type: string
          format: uuid
        target_sequence_id:
          type: integer
          minimum: 0
        reason_code:
          type: string
          minLength: 1
          maxLength: 100

    EventResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - event_id
      properties:
        status:
          type: string
          enum: [accepted]
        event_id:
          type: string
          format: uuid

    ErrorMap:
      type: object
      additionalProperties: false
      required:
        - detail
      properties:
        detail:
          type: string

    ServiceGraphProjection:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - version
        - nodes
        - edges
      properties:
        tenant_id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 0
        nodes:
          type: array
          items:
            type: string
            format: uuid
        edges:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - source
              - target
            properties:
              source:
                type: string
                format: uuid
              target:
                type: string
                format: uuid

    ResourceNodeProjection:
      type: object
      additionalProperties: false
      required:
        - node_id
        - lifecycle_state
        - cpu_cores
        - memory_gb
        - last_sequence_id
        - schema_version
      properties:
        node_id:
          type: string
          format: uuid
        lifecycle_state:
          type: string
          enum: [active, orphaned, tombstoned, frozen]
        cpu_cores:
          type: number
          minimum: 0
          maximum: 128.0
        memory_gb:
          type: number
          minimum: 0
          maximum: 1024.0
        last_sequence_id:
          type: integer
          minimum: 0
        schema_version:
          type: integer
          minimum: 1

    BackpressureTelemetry:
      type: object
      additionalProperties: false
      required:
        - status
        - utilization_rho
        - arrival_rate_hz
        - service_rate_hz
        - raw_arrival_rate_hz
        - raw_service_rate_hz
        - raw_utilization_rho
        - ema_arrival_rate_hz
        - ema_service_rate_hz
        - ema_utilization_rho
        - limit_rho
      properties:
        status:
          type: string
          enum: [healthy, overloaded]
        utilization_rho:
          type: number
        arrival_rate_hz:
          type: number
        service_rate_hz:
          type: number
        raw_arrival_rate_hz:
          type: number
          minimum: 0
        raw_service_rate_hz:
          type: number
          minimum: 0
        raw_utilization_rho:
          type: number
          minimum: 0
        ema_arrival_rate_hz:
          type: number
          minimum: 0
        ema_service_rate_hz:
          type: number
          minimum: 0
        ema_utilization_rho:
          type: number
          minimum: 0
        limit_rho:
          type: number

    ReconcilerTelemetry:
      type: object
      additionalProperties: false
      required:
        - state
        - recent_failure_count
        - opened_at
        - next_retry_at
      properties:
        state:
          type: string
          enum: [closed, open, half_open]
        recent_failure_count:
          type: integer
          minimum: 0
        opened_at:
          type: number
          nullable: true
        next_retry_at:
          type: number
          nullable: true

    GraphCentralityNode:
      type: object
      additionalProperties: false
      required:
        - node_id
        - centrality_score
        - rank
      properties:
        node_id:
          type: string
          format: uuid
        centrality_score:
          type: number
          minimum: 0
          maximum: 1
        rank:
          type: integer
          minimum: 1

    GuardrailState:
      type: object
      additionalProperties: false
      required:
        - severity
        - metric_value
        - reason
      properties:
        severity:
          type: string
        metric_value:
          type: number
        reason:
          type: string
        timestamp_utc_ms:
          type: integer
          minimum: 0

    RecentControllerEvent:
      type: object
      additionalProperties: false
      required:
        - event_id
        - event_type
        - payload
        - timestamp_utc_ms
      properties:
        event_id:
          type: string
          format: uuid
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        timestamp_utc_ms:
          type: integer
          minimum: 0

    EventSeverity:
      type: string
      enum: [INFO, WARNING, CRITICAL]

    EventSource:
      type: string
      enum: [PID, CIRCUIT_BREAKER, TOKEN_BUCKET, SYSTEM]

    TelemetryEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - timestamp
        - source
        - severity
        - type
        - message
        - metadata
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        source:
          $ref: '#/components/schemas/EventSource'
        severity:
          $ref: '#/components/schemas/EventSeverity'
        type:
          type: string
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true

    QuarantineEvent:
      type: object
      additionalProperties: false
      required:
        - quarantine_id
        - tenant_id
        - aggregate_id
        - event_id
        - sender_id
        - relation
        - incoming_vclock
        - current_vclock
        - reason
        - quarantined_at
      properties:
        quarantine_id:
          type: integer
          minimum: 1
        tenant_id:
          type: string
          format: uuid
        aggregate_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        sender_id:
          type: string
        relation:
          type: string
        incoming_vclock:
          type: object
          additionalProperties:
            type: integer
        current_vclock:
          type: object
          additionalProperties:
            type: integer
        reason:
          type: string
        quarantined_at:
          type: string
          format: date-time

    QuarantineResolveRequest:
      type: object
      additionalProperties: false
      required:
        - action
      properties:
        action:
          type: string
          enum: [DISCARD, RETRY_FORCE]

    QuarantineResolveResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - action
        - quarantine_id
        - event_id
        - aggregate_id
        - requeued
      properties:
        status:
          type: string
          enum: [resolved]
        action:
          type: string
          enum: [DISCARD, RETRY_FORCE]
        quarantine_id:
          type: integer
          minimum: 1
        event_id:
          type: string
          format: uuid
        aggregate_id:
          type: string
          format: uuid
        requeued:
          type: boolean